# docs/swagger.yaml
basePath: /
definitions:
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  models.Food:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      food_id:
        type: string
      food_image:
        type: string
      id:
        type: integer
      menu_id:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      price:
        type: number
      updatedAt:
        type: string
    required:
    - food_image
    - menu_id
    - name
    - price
    type: object
  models.Invoice:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      invoice_id:
        type: string
      order_id:
        type: string
      payment_due_date:
        type: string
      payment_method:
        type: string
      payment_status:
        type: string
      updatedAt:
        type: string
    required:
    - payment_status
    type: object
...
  models.Menu:
    properties:
      category:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      end_date:
        type: string
      id:
        type: integer
      menu_id:
        type: string
      name:
        type: string
      start_date:
        type: string
      updatedAt:
        type: string
    required:
    - category
    - menu_id
    - name
    type: object
...
host: localhost:8081
info:
  contact: {}
  description: |-
    This is a comprehensive restaurant management system API built with Go and Gin framework
    Features include user management, food menus, table reservations, orders, and invoicing
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Restaurant Management API
  version: "1.0"
paths:
  /foods:
    get:
      consumes:
      - application/json
      description: Get all foods
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all foods
      tags:
      - Foods
    post:
      consumes:
      - application/json
      description: Add a new food
      parameters:
      - description: Food object
        in: body
        name: food
        required: true
        schema:
          $ref: '#/definitions/models.Food'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Food'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Add a new food
      tags:
      - Foods
  /foods/{food_id}:
    get:
      consumes:
      - application/json
      description: Get a food by ID
      parameters:
      - description: Food ID
        in: path
        name: food_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Food'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get a food by ID
      tags:
      - Foods
    put:
      consumes:
      - application/json
      description: Update a food
      parameters:
      - description: Food ID
        in: path
        name: food_id
        required: true
        type: string
      - description: Food object
        in: body
        name: food
        required: true
        schema:
          $ref: '#/definitions/models.Food'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Food'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update a food
      tags:
      - Foods
  /invoices:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of all invoices
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all invoices
      tags:
      - Invoices
    post:
      consumes:
      - application/json
      description: Create a new invoice with the provided information
      parameters:
      - description: Invoice object
        in: body
        name: invoice
        required: true
        schema:
          $ref: '#/definitions/models.Invoice'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new invoice
      tags:
      - Invoices
  /invoices/{invoice_id}:
    get:
      consumes:
      - application/json
      description: Retrieve a specific invoice by invoice_id
      parameters:
      - description: Invoice ID
        in: path
        name: invoice_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Invoice'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get invoice by ID
      tags:
      - Invoices
    put:
      consumes:
      - application/json
      description: Update an existing invoice by invoice_id
      parameters:
      - description: Invoice ID
        in: path
        name: invoice_id
        required: true
        type: string
      - description: Invoice object
        in: body
        name: invoice
        required: true
        schema:
          $ref: '#/definitions/models.Invoice'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update an invoice
      tags:
      - Invoices
  /menus:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of all menus
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all menus
      tags:
      - Menus
    post:
      consumes:
      - application/json
      description: Create a new menu with the provided information. Requires admin
        role.
      parameters:
      - description: Menu object
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/models.Menu'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new menu (Admin only)
      tags:
      - Menus
  /menus/{menu_id}:
    get:
      consumes:
      - application/json
      description: Retrieve a specific menu by menu_id
      parameters:
      - description: Menu ID
        in: path
        name: menu_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Menu'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get menu by ID
      tags:
      - Menus
    put:
      consumes:
      - application/json
      description: Update an existing menu by menu_id. Requires admin role.
      parameters:
      - description: Menu ID
        in: path
        name: menu_id
        required: true
        type: string
      - description: Menu object
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/models.Menu'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a menu (Admin only)
      tags:
      - Menus
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...