# models/invoiceModel.go
package models

import (
	"time"

	"gorm.io/gorm"
)

type Invoice struct {
	gorm.Model
	Invoice_id       string    `json:"invoice_id"`
	Order_id         string    `json:"order_id"`
	Payment_method   *string   `json:"payment_method" validate:"eq=CARD|eq=CASH|eq="`
	Payment_status   *string   `json:"payment_status" validate:"required,eq=PENDING|eq=PAID"`
	Payment_due_date time.Time `json:"payment_due_date"`
}